use int64_t in SuffixArray

+++ test/suffix_array.cc
@@ -8,20 +8,20 @@ TEST(SuffixArray, Find) {
   SuffixArray sa("mississippi");
 
   {
-    std::vector<int> v = sa.Find("m");
+    std::vector<int64_t> v = sa.Find("m");
     ASSERT_EQ(v.size(), 1);
     ASSERT_EQ(v[0], 0);
   }
 
   {
-    std::vector<int> v = sa.Find("ss");
+    std::vector<int64_t> v = sa.Find("ss");
     ASSERT_EQ(v.size(), 2);
     ASSERT_EQ(v[0], 2);
     ASSERT_EQ(v[1], 5);
   }
 
   {
-    std::vector<int> v = sa.Find("sss");
+    std::vector<int64_t> v = sa.Find("sss");
     ASSERT_EQ(v.size(), 0);
   }
 }
